# Conflicts:
#	.idea/misc.xml

+++ app/src/main/java/me/ruyeo/newcut/data/model/Criteria.kt
@@ -3,7 +3,7 @@ package me.ruyeo.newcut.data.model
 import me.ruyeo.newcut.utils.Constants.FILTER_DISTANCE
 
 data class Criteria(
-   val longitude : Long,
-   val latitude: Long,
+   val longitude : Double,
+   val latitude: Double,
    val distance: Long = FILTER_DISTANCE
 )
